fix: refactor swap
