+++ apps/ton/src/ton/logic/swap/useSwap.ts
@@ -1,72 +1,14 @@
-import { Currency } from '@pancakeswap/ton-v2-sdk'
+import { Currency, storeSwap } from '@pancakeswap/ton-v2-sdk'
 import { storeJettonTransferMessage } from '@ton-community/assets-sdk'
-import { Address, Builder, Cell, beginCell, toNano } from '@ton/core'
+import { beginCell, toNano } from '@ton/core'
 import { SendTransactionRequest, useTonConnectUI } from '@tonconnect/ui-react'
 import { useAtomValue } from 'jotai'
 import { useCallback } from 'react'
 import { addressAtom } from 'ton/atom/addressAtom'
 import { TonContext } from 'ton/context/TonContext'
 import { Contracts } from 'ton/def/contracts.def'
-import { TON_OPCODES } from 'ton/opcodes'
 import { TonContractNames } from 'ton/ton.enums'
 import { getJettonWalletAddress, parseAddress } from 'ton/utils/address'
-
-type SwapNext = {
-  $$type: 'SwapNext'
-  tokenAddress: Address
-  minOut: bigint
-  next: Cell | null
-}
-
-type Swap = {
-  $$type: 'Swap'
-  fromUserAddress: Address
-  tokenWallet: Address
-  minOut: bigint
-  fromRealUser: Address
-  refAddress: Address | null
-  refMessageValue: bigint
-  next: SwapNext | null
-}
-
-export function storeSwap(src: Swap) {
-  return (builder: Builder) => {
-    const b0 = builder
-    b0.storeUint(TON_OPCODES.SWAP, 32)
-      .storeAddress(src.tokenWallet)
-      .storeCoins(src.minOut)
-      .storeAddress(src.fromUserAddress)
-    if (src.refAddress !== null) {
-      b0.storeBit(true).storeAddress(src.refAddress)
-    } else {
-      b0.storeBit(false)
-    }
-    b0.storeCoins(src.refMessageValue)
-
-    const b1 = new Builder()
-    if (src.next !== null && src.next !== undefined) {
-      b0.storeBit(true)
-      b1.store(storeSwapNext(src.next))
-    } else {
-      b0.storeBit(false)
-    }
-
-    b0.storeRef(b1.endCell())
-  }
-}
-
-export function storeSwapNext(src: SwapNext) {
-  return (builder: Builder) => {
-    const b0 = builder
-    b0.storeAddress(src.tokenAddress)
-    b0.storeCoins(src.minOut)
-    if (src.next !== null && src.next !== undefined) {
-      b0.storeBit(true).storeRef(src.next)
-    } else {
-      b0.storeBit(false)
-    }
-  }
-}
 
 interface SwapArgs {
   token0: Currency
@@ -124,6 +66,7 @@ export const useSwap = () => {
           {
             address: userJettonWallet0.toString(),
             // Attached TON for fees, not the amount of jettons to transfer!
+            // todo:@eric add estimate logic
             amount: toNano('0.6').toString(),
             payload: payload.toBoc().toString('base64'),
           },
