+++ packages/ton-v2-sdk/src/constants/currency/AgnosticToken.ts
