+++ packages/ton-v2-sdk/src/constants/currency/currency.ts
