+++ packages/ton-v2-sdk/src/constants/index.ts
@@ -1,4 +1,3 @@
 export * from './currency'
-export * from './Native'
-export * from './Token'
 export * from './CurrencyAmount'
+export * from './opcodes'
