+++ packages/ton-v2-sdk/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './price'
 export * from './getAddress'
 export * from './trade'
 export * from './swap'
+export * from './liquidity'
